+++ app.py
@@ -1,31 +1,43 @@
 from flask import Flask, render_template, request, redirect
-from flask_sqlalchemy import SQLAlchemy
+from models import db, Tasks
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'
-db = SQLAlchemy(app)
+db.init_app(app)
 
-class Tasks(db.Model):
-    id = db.Column(db.Integer, primary_key = True)
-    nome = db.Column(db.String(100), unique=True, nullable=False)
-    nome_cientifico = db.Column(db.String(100), unique=True, nullable=False)
-    status = db.Column(db.String(100), unique=False, nullable=False)
-    regiao = db.Column(db.String(100), unique=False, nullable=False)
-    bioma = db.Column(db.String(100), unique=False, nullable=False)
 
 @app.route('/')
 def index():
-    tasks = Tasks.query.all()
-    return render_template("index.html", tasks=tasks)
+    page = request.args.get('page', 1, type=int)
+    per_page = 10
+    pagination = Tasks.query.paginate(page=page, per_page=per_page, error_out=False)
+    tasks = pagination.items
+    return render_template("index.html", tasks=tasks, pagination=pagination)
 
 @app.route("/create", methods=["POST"])
 def create_task():
     nome = request.form["nome"]
-    new_task = Tasks(nome = nome)
-    db.session.add(new_task)
-    db.session.commit()
-    return redirect('/')
+    nome_cientifico = request.form["nome_cientifico"]
+    status = request.form["status"]
+    regiao = request.form["regiao"]
+    bioma = request.form["bioma"]
 
+    new_task = Tasks(
+        nome=nome,
+        nome_cientifico=nome_cientifico,
+        status=status,
+        regiao=regiao,
+        bioma=bioma
+    )
+
+    try:
+        db.session.add(new_task)
+        db.session.commit()
+    except Exception as e:
+        db.session.rollback()
+        return f"Erro ao salvar no banco: {e}"
+
+    return redirect('/')
 
 if __name__ == "__main__":
     with app.app_context():
